back/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
PyMySQL==1.1.0
Flask-JWT-Extended==4.5.1
Flask-Cors==4.0.0
python-dateutil==2.8.2
requests==2.31.0
Werkzeug==2.3.7
